+++ ChangeLog.md
@@ -5,6 +5,12 @@ Unreleased
 * Support for event watching: `addEventWatch` and `delEventWatch`.
 * High-level bindings now distinguish between finger down / motion / up.
   See SDL.Event.TouchFingerEvent and SDL.Event.TouchFingerMotionEvent.
+* Several event payloads now have their `Window` fields modified to use `Maybe Window`, substituting `Nothing` for null pointers.
+* High-level structure for controller button state: `ControllerButtonState`.
+* High-level structure for controller buttons: `ControllerButton`.
+* High-level structure for controller connection: `ControllerDeviceConnection`.
+* High-level structure for joystick device connection: `JoyDeviceConnection`.
+* High-level structure for joystick button state: `JoyButtonState`.
 
 2.2.0
 =====